Modularize library functions
