+++ test-site/client/src/components/QueryAll.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import Quell from '../quell-client.js'
 
 // component to get ALL data from our created DB
 const QueryAll = () => {
@@ -44,53 +45,50 @@ const QueryAll = () => {
   }
   `
 
+  // const checkStorage = (query) => {
+  //   // is query in storage?
+  //   if (sessionStorage.getItem(query)) return true
+  //   return false
+  // }
 
+  // const serveFromCache = (query) => {
+  //   // return from storage and update state
+  //   console.log('Serving from cache')
+  //   return sessionStorage.getItem(query)
+  // }
 
-  //do check on cache for query ? serve from there : fetch(//query), store, display
+  // const fetchAndServe = (query, endpoint = '/graphql') => {
+  //   // return from fetch and update state
+  //   fetch(endpoint, {
+  //     method: 'POST',
+  //     headers: {
+  //       'Content-Type': 'application/json'
+  //     },
+  //     body: JSON.stringify({ query: query })
+  //   })
+  //     .then(res => res.json())
+  //     .then(res => {
+  //       const responseData = JSON.stringify(res.data);
+  //       console.log('Saving to cache')
+  //       sessionStorage.setItem(query, responseData);
 
-  const checkStorage = (query) => {
-    // is query in storage?
-    if (sessionStorage.getItem(query)) return true
-    return false
-  }
+  //       return responseData;
+  //     })
+  //     .catch(err => console.log(err))
+  // }
 
-  const serveFromCache = (query) => {
-    // return from storage and update state
-    console.log('Serving from cache')
-    return sessionStorage.getItem(query)
-  }
+  // const quellFetch = (query) => {
+  //   // check if full query is in cache, if so serve result from cache
+  //   if (checkStorage(query)) return serveFromCache(query)
 
-  const fetchAndServe = (query, endpoint = '/graphql') => {
-    // return from fetch and update state
-    fetch(endpoint, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ query: query })
-    })
-      .then(res => res.json())
-      .then(res => {
-        const responseData = JSON.stringify(res.data);
-        console.log('Saving to cache')
-        sessionStorage.setItem(query, responseData);
-
-        return responseData;
-      })
-      .catch(err => console.log(err))
-  }
-
-  const quellFetch = (query) => {
-    // check if full query is in cache, if so serve result from cache
-    if (checkStorage(query)) return serveFromCache(query)
-
-    // query not found in cache, fetch data from server and return data
-    return fetchAndServe(query)
-  }
+  //   // query not found in cache, fetch data from server and return data
+  //   return fetchAndServe(query)
+  // }
 
   const handleClick = () => {
     // run quellFetch() to receive result
-    const displayResults = quellFetch(queryCall)
+    const displayResults = Quell.quellFetch(queryCall)
+    console.log('handleClick for displayResults', displayResults)
     // update state to display results
     setQueryResponse(displayResults);
   }
